Comments and iter.list
Added comments to all of the new code.
Also, changed how static_list objects are instantiated such that the
static_iterator does all of the work. This is done by having static_list
take in the object directly and calling the iter function. Later, this
function will have to call the iter.static function to ensure that the
function returns a static_iterator.