Stopwatch
stopwatch.js provides a class that has three main functions
start - Clocks the current time using either performance.now or Date.now
stop- Clocks the current times using either performance.now or Date.now
span - Returns the difference between start and stop times and
multiplies it by StopWatch.Milliseconds (one) or the value provided.
Also, provides a Clock function that will create a StopWatch instance
and use it to calculate the time to run the function provided.

It is important to note that a time of zero is very difficult to create.
So, either the start/stop was never called before calling span or the
data was poor.